Fix problem with private members.
The existing code encodes all the properties of a data class even ones not
in the primary constructor but only ever uses the primary constructor.
Any private properties cause this code to fail as we do not have access to
them (the isAccessible trick doesn't work on private properties that do not
have a backing store).

This changes the code to only encode the properties of the primary constructor's
parameters. As data classes only have property parameters in the primary
constructor, this is sufficient.

As part of reading the values the isAccessible flag is adjusted to make the
values accessible and now the logic resets the isAccessible flag to its original
value after processing.

Also added a check for IllegalAccessException that displays the parameter that
caused the problem to help with debugging in the future. In theory this won't
happen, but if something changes in how properties are handled in the future
this should help identify the issue.