Updating comparison table for ktoml project
I love that table and marketing ideas to promote your project with this table. Very good job (no sarcasm)! 
If you will write an MPP-based parser for TOML - I would definitely like to use it in [ktoml](https://github.com/akuleshov7/ktoml). 

Also my suggestion is to mention on which phase the decoding/encoding process is done: on compile time with Kotlin compiler plugin or on the runtime with reflection.